Add files via upload
This program processes the data in Inventory.txt using instances of the Inventory class. Each time the
customer purchases or returns an item, the program uses an instance of TransactionItem to store that information. When
a customer completes a transaction (a customer may both purchase and return items in the same transaction), the program
will display an invoice to the console, and create a new text file (UpdatedInventory.txt) with all inventory information
and updated quantities.